Jadid/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
clamd==1.0.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.0.0
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.15
django-allauth==0.61.1
django-bootstrap-sidebar==1.2.3
django-extensions==3.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
django-clamd==1.0.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwcrypto==1.5.4
kombu==5.3.2
ldap3==2.9.1
msgpack==1.0.7
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pillow==10.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-keycloak==3.9.1
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.5.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.9
xlrd==2.0.1
XlsxWriter==3.0.3
zope.interface==6.1

